Add new containsLocation() that takes latitude and longitude instead of LatLng (#331)
In our project, we use different data structure to represent coordinates rather than LatLng. As a result, in order to utilize `containsLocation()`, we have to convert into LatLng which causes an allocation. So if we call it inside a loop through a list of coordinates, that will cause a lot more allocations.