Demo clustering with ViewModel (#506)
* Differentiate between initial start and configuration change restore

Only move the camera to the starting location when not restoring

* Add removeItems() to ClusterManager and Algorithm

Avoid locking and unlocking the algorithm in a tight loop when removing many cluster items

* Move algorithm lock from ClusterManager to Algorithm

Allow the algorithm to handle its own locking, so it can be reused in another ClusterManager instance. Also allow adding/removing items with proper locking outside of ClusterManager. For best ViewModel support.