fix: Respect setMinClusterSize() correctly when clustering
Previously, when `setMinClusterSize(4)` was called, clusters of size 4 would still be rendered as individual markers instead of clusters (only clusters of size 5 or more would be rendered as individual cluster markers).

With this change, now `setMinClusterSize(4)` will result in clusters of size 4 being rendered as clusters instead of individual markers.

Closes #696